refactor: reorganize recommendations structure with component.tsx pattern
Reorganize the Recommendations directory to follow consistent component.tsx naming:
- Move components out of nested /components directory structure
- Rename all main component files to component.tsx
- Add index.ts barrel exports for clean imports
- Update all import paths throughout codebase
- Maintain existing functionality and test coverage

Directory structure changes:
- RecommendationCard: RecommendationCard.tsx → component.tsx
- RecommendationFilters: RecommendationFilters.tsx → component.tsx
- RecommendationsSection: RecommendationsSection.tsx → component.tsx
- Add index.ts files for each component directory

Updated imports in layout, search, and shared components to use new paths.
