+++ src/domains/layout/Recommendations/RecommendationCard/RecommendationCard.skills.test.tsx
@@ -1,7 +1,8 @@
 import { render, screen, within } from '@testing-library/react';
-import { RecommendationCard } from './RecommendationCard';
-import type { Recommendation } from '../../../../../data/recommendations';
-import type { Filter, SkillFilter } from '../../types/filtering';
+import { RecommendationCard } from './component';
+import type { Recommendation } from '../../../../data/recommendations';
+import type { Filter, SkillFilter } from '../types/filtering';
+
 
 // Mock recommendation with diverse skills across actual categories from the system
 const mockRecommendationWithCategorizedSkills: Recommendation = {
