+++ src/domains/layout/Recommendations/RecommendationCard/component.module.css
