+++ src/domains/layout/Recommendations/RecommendationCard/component.tsx
@@ -1,13 +1,14 @@
 import React, { useEffect, useState } from 'react';
-import type { Recommendation } from '../../../../../data/recommendations';
-import type { Filter, SortOption, SortOrder } from '../../types/filtering';
-import { trackRecommendationEvent } from '../../../../analytics/utils/eventHelpers';
+import type { Recommendation } from '../../../../data/recommendations';
+import { trackRecommendationEvent } from '../../../analytics/utils/eventHelpers';
 
-import { HighlightedText } from '../../../../search/components/HighlightedText/HighlightedText';
-import { SkillTags } from '../../../../../shared/components/ui/SkillTags';
-import styles from './RecommendationCard.module.css';
-import { useHighlightedTerms } from '../../hooks/useHighlightedTerms';
-import type { SkillCategory } from '../../../../../data/skills';
+import { HighlightedText } from '../../../search/components/HighlightedText/HighlightedText';
+import { SkillTags } from '../../../../shared/components/ui/SkillTags';
+import styles from './component.module.css';
+
+import type { SkillCategory } from '../../../../data/skills';
+import type { Filter, SortOption, SortOrder } from '../types/filtering';
+import { useHighlightedTerms } from '../hooks';
 
 interface RecommendationCardProps {
   recommendation: Recommendation;
