+++ src/domains/layout/Recommendations/RecommendationCard/index.ts
@@ -1,0 +1,1 @@
+export { RecommendationCard } from './component';
