+++ src/domains/layout/Recommendations/RecommendationFilters/component.test.tsx
@@ -1,10 +1,10 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { RecommendationFilters } from './RecommendationFilters';
-import type { UseFiltersReturn } from '../../types/filtering';
-import type { Recommendation } from '../../../../../data/recommendations';
-import type { SkillCategory } from '../../../../../data/skills';
+import { RecommendationFilters } from './component';
+import type { UseFiltersReturn } from '../types/filtering';
+import type { Recommendation } from '../../../../data/recommendations';
+import type { SkillCategory } from '../../../../data/skills';
 
 // Mock the useRecommendationFilters hook behavior
 const mockFilters: UseFiltersReturn = {
