+++ src/domains/layout/Recommendations/RecommendationFilters/index.ts
@@ -1,0 +1,1 @@
+export { RecommendationFilters } from './component';
