+++ src/domains/layout/Recommendations/component.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import { useQuery } from '@apollo/client';
 import { Users } from 'lucide-react';
-import { GET_RECOMMENDATIONS } from '../../../../apollo/queries';
-import { RecommendationCard } from './RecommendationCard/RecommendationCard';
-import { RecommendationFilters } from './RecommendationFilters/RecommendationFilters';
-import { SearchBar } from '../../../search/components/SearchBar/SearchBar';
-import { SortingControls } from '../../../../shared/components/ui/SortingControls/SortingControls';
-import { Pagination } from '../../../../shared/components/ui/Pagination/Pagination';
-import { useRecommendationFilters } from '../hooks/useRecommendationFilters';
-import type { Recommendation } from '../../../../data/recommendations';
-import styles from '../../component.module.css';
+import { GET_RECOMMENDATIONS } from '../../../apollo/queries';
+import { RecommendationCard } from './RecommendationCard';
+import { RecommendationFilters } from './RecommendationFilters';
+import { SearchBar } from '../../search/components/SearchBar/SearchBar';
+import { SortingControls } from '../../../shared/components/ui/SortingControls/SortingControls';
+import { Pagination } from '../../../shared/components/ui/Pagination/Pagination';
+import { useRecommendationFilters } from './hooks/useRecommendationFilters';
+import type { Recommendation } from '../../../data/recommendations';
+import styles from '../component.module.css';
 
 export const RecommendationsSection: React.FC = () => {
   const { loading, error, data } = useQuery(GET_RECOMMENDATIONS);
