+++ src/domains/layout/Recommendations/index.ts
@@ -1,0 +1,1 @@
+export { RecommendationsSection } from './component';
