+++ src/shared/components/ui/SkillTags.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from 'react';
 import type { Recommendation } from '../../../data/recommendations';
 import { getSkillCategory, type SkillCategory } from '../../../data/skills';
-import styles from '../../../domains/layout/recommendations/components/RecommendationCard/RecommendationCard.module.css';
+import styles from '../../../domains/layout/recommendations/components/RecommendationCard/component.module.css';
 
 interface SkillTagsProps {
   recommendation: Recommendation;
